refactor: RequestUtils.java to RequestUtil.java and refactor repository folder
