+++ src/main/java/hangouh/me/medi/link/v1/controllers/PatientController.java
@@ -5,8 +5,8 @@ import hangouh.me.medi.link.v1.DTO.requests.PatientFilterDTO;
 import hangouh.me.medi.link.v1.DTO.responses.ResponseDTO;
 import hangouh.me.medi.link.v1.DTO.responses.ResponsePageDTO;
 import hangouh.me.medi.link.v1.models.Patient;
-import hangouh.me.medi.link.v1.repository.PatientRepository;
-import hangouh.me.medi.link.v1.utls.RequestUtils;
+import hangouh.me.medi.link.v1.repositories.PatientRepository;
+import hangouh.me.medi.link.v1.utils.RequestUtil;
 import jakarta.validation.Valid;
 import java.util.*;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,7 +31,7 @@ public class PatientController {
 
   @GetMapping("/patients")
   public ResponseEntity<ResponsePageDTO<Patient>> getAll(@Valid PatientFilterDTO dto) {
-    Sort sort = RequestUtils.buildSort(dto.getSortBy());
+    Sort sort = RequestUtil.buildSort(dto.getSortBy());
     Pageable pageable = PageRequest.of(dto.getPage(), dto.getSize(), sort);
     Page<Patient> patients = patientRepository.findByFilters(dto, pageable);
     return ResponseEntity.ok(new ResponsePageDTO<>(HttpStatus.OK, "", patients));
